backend/requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
Pillow
requests
python-dotenv
openai